requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
appnope==0.1.4
asttokens==2.4.1
-e git+https://github.com/thobs-10/ml-mid-term-project.git@eaf9cdfa1ac34f9e0e61a1de2f60e70329d94859#egg=Bike_Sharing_Demand_Predictions
catboost==1.2.7
click==8.1.7
cloudpickle==3.1.0
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.8
decorator==5.1.1
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.115.5
fonttools==4.54.1
future==1.0.0
graphviz==0.20.3
h11==0.14.0
hyperopt==0.2.7
idna==3.10
imbalanced-learn==0.12.4
ipykernel==6.29.5
ipython==8.29.0
jedi==0.19.1
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
lightgbm==4.5.0
loguru==0.7.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
networkx==3.4.2
numpy==1.23.5
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
plotly==5.24.1
prompt_toolkit==3.0.48
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
py4j==0.10.9.7
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pyzmq==26.2.0
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.41.3
tenacity==9.0.0
threadpoolctl==3.5.0
tornado==6.4.1
tqdm==4.67.0
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.32.1
wcwidth==0.2.13
xgboost==2.1.2
